refactor(community): restructure hierarchy of card in my-information
