+++ apps/community/src/features/profile/components/editable-my-profile-card.css.ts
@@ -1,0 +1,22 @@
+import { style } from '@vanilla-extract/css'
+
+import { themeVars } from '@aics-client/design-system/styles'
+
+const fieldTitle = style({
+  width: '6rem',
+  fontSize: themeVars.fontSize.lg,
+  fontWeight: themeVars.fontWeight.semibold,
+})
+
+const editFieldWrapper = style({
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'space-between',
+})
+
+const editField = style({
+  margin: '1.5rem 0',
+  width: '20rem',
+})
+
+export { fieldTitle, editFieldWrapper, editField }
