+++ apps/community/src/features/profile/components/editable-my-profile-card.tsx
@@ -1,17 +1,47 @@
 'use client'
 
-import { zodResolver } from '@hookform/resolvers/zod'
-import { useForm } from 'react-hook-form'
+import type { UseFormRegisterReturn } from 'react-hook-form'
 import type { z } from 'zod'
 
-import { Button } from '@aics-client/design-system'
+import { Button, Input } from '@aics-client/design-system'
 
+import * as styles from '~/features/profile/components/editable-my-profile-card.css'
 import { MyInfoCard } from '~/features/profile/components/my-info-card'
 import { editMyInfoSchema } from '~/features/profile/schemas/edit-my-info-schema'
 import { useEditProfile } from '~/features/profile/services/use-edit-profile.mutation'
 import type { UserEditableDetail } from '~/features/profile/types/profile'
+import { useZodForm } from '~/shared/hooks/use-zod-form'
 
-type FormValues = z.infer<typeof editMyInfoSchema>
+type EditProfileFormValues = z.infer<typeof editMyInfoSchema>
+
+interface MyInfoEditableFieldProps {
+  title: string
+  value: string
+  register?: UseFormRegisterReturn
+  error?: string
+}
+
+function MyInfoEditableField({
+  title,
+  value,
+  register,
+  error,
+}: MyInfoEditableFieldProps) {
+  return (
+    <section className={styles.editFieldWrapper}>
+      <div>
+        <h3 className={styles.fieldTitle}>{title}</h3>
+        <Input
+          className={styles.editField}
+          type="text"
+          defaultValue={value}
+          message={error}
+          {...register}
+        />
+      </div>
+    </section>
+  )
+}
 
 function EditableMyProfileCard({
   initialData,
@@ -30,22 +60,21 @@ function EditableMyProfileCard({
     register,
     handleSubmit,
     formState: { errors, isDirty, isSubmitted },
-  } = useForm<FormValues>({
-    resolver: zodResolver(editMyInfoSchema),
-    defaultValues,
+  } = useZodForm({
+    schema: editMyInfoSchema,
+    defaultValues: defaultValues,
     mode: 'onChange',
   })
 
-  const handleFormSubmit = (formData: FormValues) => {
+  const handleFormSubmit = (formData: EditProfileFormValues) => {
     mutation.mutate(formData)
-    alert('회원 정보가 수정되었습니다.')
   }
 
   return (
     <MyInfoCard title="기본 정보">
       <form onSubmit={handleSubmit(handleFormSubmit)}>
         {initialData.map((detail) => (
-          <MyInfoCard.EditableField
+          <MyInfoEditableField
             key={detail.field}
             title={detail.title}
             value={detail.value}
