+++ apps/community/src/features/profile/components/my-info-card.css.ts
@@ -25,33 +25,4 @@ const cardContent = style({
   gap: themeVars.spacing.lg,
 })
 
-const field = style({
-  margin: '1.5rem 0',
-})
-
-const fieldTitle = style({
-  width: '6rem',
-  fontSize: themeVars.fontSize.lg,
-  fontWeight: themeVars.fontWeight.semibold,
-})
-
-const editFieldWrapper = style({
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'space-between',
-})
-
-const editField = style({
-  margin: '1.5rem 0',
-  width: '20rem',
-})
-
-export {
-  cardWrapper,
-  cardTitle,
-  cardContent,
-  field,
-  fieldTitle,
-  editFieldWrapper,
-  editField,
-}
+export { cardWrapper, cardTitle, cardContent }
